Add functionality to link to rover
∆ Controller can now link to a mars rover with #link_to_rover. The
controller must be able to link to a rover in order to land it. In the
future the controller should have the ability to switch which rover it is
commanding, this leaves allowence for that.

∆ When controller impliments #link_to_rover, a rover is created. This is
the basic functionality to pass the test of having a rover created. In
the future, this functionality will have to be looked at again when the
controller is required to link to an existing rover.

∆ Create the MarsRover class with a name on initialization. When linking
to a named rover, the rover needs to exists as a class with a name
attribute.