Add Plateau to Controller
∆ Add Plateau to NasaController on initialization. By injecting a
plateau in this way, the controller is dependant on a Plateau, but is
still very changable. This reduces the costs of the dependency. I
believe this dependency is fine as in it's current state, thier is no
reason you would have a nasa_controller without a plateau, and there is
no forseeable future that would alter this.